deleteing projects works

+++ server/controllers/deleteController.js
@@ -22,40 +22,53 @@ const removeProjectAccess = async (req, res) => {
   res.sendStatus(200);
 };
 const removeProject = async (req, res) => {
+  console.log("------Deleting Project ----------");
+  console.log("req_project_body: ", req);
+
   const { project_id } = req.body;
-  await sequelize.query("DELETE FROM projects WHERE project_id=?", {
-    replacements: [project_id],
-  });
-  let projectGroup = await sequelize.query(
-    "DELETE FROM project_group WHERE project_group_id=? RETURNING project_group_id",
-    {
+  if (project_id && project_id > 0) {
+    let group_id = await sequelize.query(
+      "SELECT * FROM project_group_access WHERE project_group_id=(SELECT project_group_id FROM project_group WHERE project_id=?)",
+      {
+        replacements: [project_id],
+      });
+    let project_group_id = group_id[0][0].project_group_id
+    console.log("project_id: ", project_id);
+    console.log("projectGroupID: ", project_group_id);
+    console.log(group_id[0][0])
+    await sequelize.query(
+      "DELETE FROM project_group_access WHERE project_group_id=?",
+      {
+        replacements: [project_group_id],
+      });
+    await sequelize.query(
+      "DELETE FROM project_group WHERE project_group_id=? RETURNING project_group_id",
+      {
+        replacements: [project_group_id],
+      });
+    await sequelize.query("DELETE FROM project_spells WHERE project_id=?", {
       replacements: [project_id],
-    }
-  );
-  let projectGroupID = projectGroup[0][0].project_group_id;
-  await sequelize.query(
-    "DELETE FROM project_group_access WHERE project_group_id=?",
-    {
-      replacements: [projectGroupID],
-    }
-  );
-  await sequelize.query("DELETE FROM project_spells WHERE project_id=?", {
-    replacements: [project_id],
-  });
-  await sequelize.query("DELETE FROM project_feats WHERE project_id=?", {
-    replacements: [project_id],
-  });
-  await sequelize.query("DELETE FROM project_abilities WHERE project_id=?", {
-    replacements: [project_id],
-  });
-  await sequelize.query("DELETE FROM project_classes WHERE project_id=?", {
-    replacements: [project_id],
-  });
-  await sequelize.query("DELETE FROM project_subclasses WHERE project_id=?", {
-    replacements: [project_id],
-  });
-  res.sendStatus(200);
-};
+    });
+    await sequelize.query("DELETE FROM project_feats WHERE project_id=?", {
+      replacements: [project_id],
+    });
+    await sequelize.query("DELETE FROM project_abilities WHERE project_id=?", {
+      replacements: [project_id],
+    });
+    await sequelize.query("DELETE FROM project_classes WHERE project_id=?", {
+      replacements: [project_id],
+    });
+    await sequelize.query("DELETE FROM project_subclasses WHERE project_id=?", {
+      replacements: [project_id],
+    });
+    await sequelize.query("DELETE FROM projects WHERE project_id=?", {
+      replacements: [project_id],
+    });
+    res.sendStatus(200);
+  } else {
+    res.sendStatus(400)
+  }
+}
 const removeClass = async (req, res) => {
   const { class_id } = req.body;
   await sequelize.query("DELETE FROM classes WHERE class_id=?", {
@@ -99,4 +112,4 @@ module.exports = {
   removeAbility,
   removeSpell,
   removeFeat,
-};
+}
